FEATURE: el6: Update to pywbem version 0.11.0
The version of python-setuptools shipped with CentOS 6 is too old to
build the package (0.6.10 but requires at least 12.0), so we use a
more recent one during the build process. We still need the standard
python-setuptools package to build the more recent setuptools module,
hence the build dependency.

The python-ply dependency is available from the EPEL 6 repository.
Other dependencies are available in the standard CentOS 6 repository.